Junit tests for zOS Console and consistent package names for batch manager
