--- galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosbatch.zosmf.manager/src/main/java/dev/galasa/zosbatch/zosmf/internal/package-info.java
@@ -1,9 +1,0 @@
-/*
- * Licensed Materials - Property of IBM
- * 
- * (c) Copyright IBM Corp. 2019.
- */
-/**
- * zOS Batch Manager - zOSMF Internal Implementation
- */
-package dev.galasa.zosbatch.zosmf.internal;
