--- galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosbatch.zosmf.manager/src/main/java/dev/galasa/zosbatch/zosmf/internal/properties/package-info.java
@@ -1,17 +1,0 @@
-/*
- * Licensed Materials - Property of IBM
- * 
- * (c) Copyright IBM Corp. 2019.
- */
-/**
- * zOS Batch zOSMF Manager - CPS Properties
- * 
- * <p>
- * The following properties are used by the zOS Batch zOSMF Manager:<br>
- * {@code zosbatch.jobname.[imageid].prefix} <br>
- * {@code zosbatch.batchjob.[imageid].timeout} <br>
- * {@code zosbatch.batchjob.[imageid].request.retry <br>
- * {@code zosbatch.batchjob.[imageid].use.sysaff} <br>
- * </p>
- */
-package dev.galasa.zosbatch.zosmf.internal.properties;
