+++ galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosbatch.zosmf.manager/src/main/java/dev/galasa/zosbatch/zosmf/manager/internal/properties/ZosBatchZosmfPropertiesSingleton.java
@@ -3,20 +3,20 @@
  * 
  * (c) Copyright IBM Corp. 2019.
  */
-package dev.galasa.zosconsole.zosmf.manager.internal.properties;
+package dev.galasa.zosbatch.zosmf.manager.internal.properties;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 
-import dev.galasa.zosconsole.ZosConsoleManagerException;
+import dev.galasa.zosbatch.ZosBatchManagerException;
 import dev.galasa.framework.spi.IConfigurationPropertyStoreService;
 
-@Component(service=ZosConsoleZosmfPropertiesSingleton.class, immediate=true)
-public class ZosConsoleZosmfPropertiesSingleton {
+@Component(service=ZosBatchZosmfPropertiesSingleton.class, immediate=true)
+public class ZosBatchZosmfPropertiesSingleton {
     
-    private static ZosConsoleZosmfPropertiesSingleton singletonInstance;
-    private static void setInstance(ZosConsoleZosmfPropertiesSingleton instance) {
+    private static ZosBatchZosmfPropertiesSingleton singletonInstance;
+    private static void setInstance(ZosBatchZosmfPropertiesSingleton instance) {
         singletonInstance = instance;
     }
     
@@ -32,20 +32,20 @@ public class ZosConsoleZosmfPropertiesSingleton {
         setInstance(null);
     }
     
-    public static IConfigurationPropertyStoreService cps() throws ZosConsoleManagerException {
+    public static IConfigurationPropertyStoreService cps() throws ZosBatchManagerException {
         if (singletonInstance != null) {
             return singletonInstance.cps;
         }
         
-        throw new ZosConsoleManagerException("Attempt to access manager CPS before it has been initialised");
+        throw new ZosBatchManagerException("Attempt to access manager CPS before it has been initialised");
     }
     
-    public static void setCps(IConfigurationPropertyStoreService cps) throws ZosConsoleManagerException {
+    public static void setCps(IConfigurationPropertyStoreService cps) throws ZosBatchManagerException {
         if (singletonInstance != null) {
             singletonInstance.cps = cps;
             return;
         }
         
-        throw new ZosConsoleManagerException("Attempt to set manager CPS before instance created");
+        throw new ZosBatchManagerException("Attempt to set manager CPS before instance created");
     }
 }
