+++ galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosconsole.zosmf.manager/src/main/java/dev/galasa/zosconsole/zosmf/manager/internal/ZosConsoleImpl.java
@@ -22,29 +22,29 @@ import dev.galasa.zos.IZosImage;
  */
 public class ZosConsoleImpl implements IZosConsole {
 
-	private List<ZosConsoleCommandImpl> zosConsoleCommands = new ArrayList<>();
-	private IZosImage image;
-	
-	public ZosConsoleImpl(IZosImage image) {
-		this.image = image;
-	}
+    private List<ZosConsoleCommandImpl> zosConsoleCommands = new ArrayList<>();
+    private IZosImage image;
+    
+    public ZosConsoleImpl(IZosImage image) {
+        this.image = image;
+    }
 
-	@Override
-	public @NotNull IZosConsoleCommand issueCommand(@NotNull String command) throws ZosConsoleException {
-		return issueCommand(command, null);
-	}
+    @Override
+    public @NotNull IZosConsoleCommand issueCommand(@NotNull String command) throws ZosConsoleException {
+        return issueCommand(command, null);
+    }
 
-	@Override
-	public @NotNull IZosConsoleCommand issueCommand(@NotNull String command, String consoleName) throws ZosConsoleException {
-		ZosConsoleCommandImpl zosConsoleCommand;
-		
-		try {
-			zosConsoleCommand = new ZosConsoleCommandImpl(command, consoleName, this.image);
-			this.zosConsoleCommands.add(zosConsoleCommand);
-		} catch (ZosConsoleManagerException e) {
-			throw new ZosConsoleException("Unable to issue console command", e);
-		}
-		
-		return zosConsoleCommand.issueCommand();
-	}
+    @Override
+    public @NotNull IZosConsoleCommand issueCommand(@NotNull String command, String consoleName) throws ZosConsoleException {
+        ZosConsoleCommandImpl zosConsoleCommand;
+        
+        try {
+            zosConsoleCommand = new ZosConsoleCommandImpl(command, consoleName, this.image);
+            this.zosConsoleCommands.add(zosConsoleCommand);
+        } catch (ZosConsoleManagerException e) {
+            throw new ZosConsoleException("Unable to issue console command", e);
+        }
+        
+        return zosConsoleCommand.issueCommand();
+    }
 }
