+++ galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosconsole.zosmf.manager/src/main/java/dev/galasa/zosconsole/zosmf/manager/internal/ZosConsoleManagerImpl.java
@@ -38,71 +38,71 @@ import dev.galasa.zosconsole.zosmf.manager.internal.properties.ZosConsoleZosmfPr
  */
 @Component(service = { IManager.class })
 public class ZosConsoleManagerImpl extends AbstractManager {
-	protected static final String NAMESPACE = "zosconsole";
+    protected static final String NAMESPACE = "zosconsole";
 
-	protected static IZosManagerSpi zosManager;
-	public static void setZosManager(IZosManagerSpi zosManager) {
-		ZosConsoleManagerImpl.zosManager = zosManager;
-	}
-	
-	protected static IZosmfManagerSpi zosmfManager;
-	public static void setZosmfManager(IZosmfManagerSpi zosmfManager) {
-		ZosConsoleManagerImpl.zosmfManager = zosmfManager;
-	}
+    protected static IZosManagerSpi zosManager;
+    public static void setZosManager(IZosManagerSpi zosManager) {
+        ZosConsoleManagerImpl.zosManager = zosManager;
+    }
+    
+    protected static IZosmfManagerSpi zosmfManager;
+    public static void setZosmfManager(IZosmfManagerSpi zosmfManager) {
+        ZosConsoleManagerImpl.zosmfManager = zosmfManager;
+    }
 
-	private final HashMap<String, ZosConsoleImpl> taggedZosConsoles = new HashMap<>();
-	
-	/* (non-Javadoc)
-	 * @see dev.galasa.framework.spi.AbstractManager#initialise(dev.galasa.framework.spi.IFramework, java.util.List, java.util.List, java.lang.Class)
-	 */
-	@Override
-	public void initialise(@NotNull IFramework framework, @NotNull List<IManager> allManagers,
-			@NotNull List<IManager> activeManagers, @NotNull Class<?> testClass) throws ManagerException {
-		super.initialise(framework, allManagers, activeManagers, testClass);
-		try {
-			ZosConsoleZosmfPropertiesSingleton.setCps(framework.getConfigurationPropertyService(NAMESPACE));
-		} catch (ConfigurationPropertyStoreException e) {
-			throw new ZosConsoleManagerException("Unable to request framework services", e);
-		}
+    private final HashMap<String, ZosConsoleImpl> taggedZosConsoles = new HashMap<>();
+    
+    /* (non-Javadoc)
+     * @see dev.galasa.framework.spi.AbstractManager#initialise(dev.galasa.framework.spi.IFramework, java.util.List, java.util.List, java.lang.Class)
+     */
+    @Override
+    public void initialise(@NotNull IFramework framework, @NotNull List<IManager> allManagers,
+            @NotNull List<IManager> activeManagers, @NotNull Class<?> testClass) throws ManagerException {
+        super.initialise(framework, allManagers, activeManagers, testClass);
+        try {
+            ZosConsoleZosmfPropertiesSingleton.setCps(framework.getConfigurationPropertyService(NAMESPACE));
+        } catch (ConfigurationPropertyStoreException e) {
+            throw new ZosConsoleManagerException("Unable to request framework services", e);
+        }
 
-		//*** Check to see if any of our annotations are present in the test class
-		//*** If there is,  we need to activate
-		List<AnnotatedField> ourFields = findAnnotatedFields(ZosConsoleField.class);
-		if (!ourFields.isEmpty()) {
-			youAreRequired(allManagers, activeManagers);
-		}
-	}
-	
+        //*** Check to see if any of our annotations are present in the test class
+        //*** If there is,  we need to activate
+        List<AnnotatedField> ourFields = findAnnotatedFields(ZosConsoleField.class);
+        if (!ourFields.isEmpty()) {
+            youAreRequired(allManagers, activeManagers);
+        }
+    }
+    
 
-	/* (non-Javadoc)
-	 * @see dev.galasa.framework.spi.AbstractManager#provisionGenerate()
-	 */
-	@Override
-	public void provisionGenerate() throws ManagerException, ResourceUnavailableException {
-		generateAnnotatedFields(ZosConsoleField.class);
-	}
+    /* (non-Javadoc)
+     * @see dev.galasa.framework.spi.AbstractManager#provisionGenerate()
+     */
+    @Override
+    public void provisionGenerate() throws ManagerException, ResourceUnavailableException {
+        generateAnnotatedFields(ZosConsoleField.class);
+    }
 
 
-	/* (non-Javadoc)
-	 * @see dev.galasa.framework.spi.AbstractManager#youAreRequired()
-	 */
-	@Override
-	public void youAreRequired(@NotNull List<IManager> allManagers, @NotNull List<IManager> activeManagers)
-			throws ManagerException {
-		if (activeManagers.contains(this)) {
-			return;
-		}
+    /* (non-Javadoc)
+     * @see dev.galasa.framework.spi.AbstractManager#youAreRequired()
+     */
+    @Override
+    public void youAreRequired(@NotNull List<IManager> allManagers, @NotNull List<IManager> activeManagers)
+            throws ManagerException {
+        if (activeManagers.contains(this)) {
+            return;
+        }
 
-		activeManagers.add(this);
-		setZosManager(addDependentManager(allManagers, activeManagers, IZosManagerSpi.class));
-		if (zosManager == null) {
-			throw new ZosConsoleManagerException("The zOS Manager is not available");
-		}
-		setZosmfManager(addDependentManager(allManagers, activeManagers, IZosmfManagerSpi.class));
-		if (zosmfManager == null) {
-			throw new ZosConsoleManagerException("The zOSMF Manager is not available");
-		}
-	}
+        activeManagers.add(this);
+        setZosManager(addDependentManager(allManagers, activeManagers, IZosManagerSpi.class));
+        if (zosManager == null) {
+            throw new ZosConsoleManagerException("The zOS Manager is not available");
+        }
+        setZosmfManager(addDependentManager(allManagers, activeManagers, IZosmfManagerSpi.class));
+        if (zosmfManager == null) {
+            throw new ZosConsoleManagerException("The zOSMF Manager is not available");
+        }
+    }
 
     /*
      * (non-Javadoc)
@@ -113,26 +113,26 @@ public class ZosConsoleManagerImpl extends AbstractManager {
      */
     @Override
     public boolean areYouProvisionalDependentOn(@NotNull IManager otherManager) {
-    	return otherManager instanceof IZosManagerSpi ||
-    		   otherManager instanceof IZosmfManagerSpi;
+        return otherManager instanceof IZosManagerSpi ||
+               otherManager instanceof IZosmfManagerSpi;
     }
 
     @GenerateAnnotatedField(annotation=ZosConsole.class)
-	public IZosConsole generateZosConsole(Field field, List<Annotation> annotations) throws ZosManagerException {
-		ZosConsole annotationZosConsole = field.getAnnotation(ZosConsole.class);
+    public IZosConsole generateZosConsole(Field field, List<Annotation> annotations) throws ZosManagerException {
+        ZosConsole annotationZosConsole = field.getAnnotation(ZosConsole.class);
 
-		//*** Default the tag to primary
-		String tag = defaultString(annotationZosConsole.imageTag(), "primary");
+        //*** Default the tag to primary
+        String tag = defaultString(annotationZosConsole.imageTag(), "primary");
 
-		//*** Have we already generated this tag
-		if (this.taggedZosConsoles.containsKey(tag)) {
-			return this.taggedZosConsoles.get(tag);
-		}
+        //*** Have we already generated this tag
+        if (this.taggedZosConsoles.containsKey(tag)) {
+            return this.taggedZosConsoles.get(tag);
+        }
 
-		IZosImage image = zosManager.getImageForTag(tag);
-		IZosConsole zosConsole = new ZosConsoleImpl(image);
-		this.taggedZosConsoles.put(tag, (ZosConsoleImpl) zosConsole);
-		
-		return zosConsole;
-	}
+        IZosImage image = zosManager.getImageForTag(tag);
+        IZosConsole zosConsole = new ZosConsoleImpl(image);
+        this.taggedZosConsoles.put(tag, (ZosConsoleImpl) zosConsole);
+        
+        return zosConsole;
+    }
 }
