+++ galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosconsole.zosmf.manager/src/main/java/dev/galasa/zosconsole/zosmf/manager/internal/package-info.java
@@ -1,3 +1,8 @@
+/*
+ * Licensed Materials - Property of IBM
+ * 
+ * (c) Copyright IBM Corp. 2019.
+ */
 /**
  * zOS Console Manager - zOSMF Internal Implementation
  */
